Split capability authorization state helpers
